# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Coconut API",
    "version": "1.0.0",
    "description": "API documentation for the Coconut application"
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/check/auth-status": {
      "get": {
        "summary": "Validate the authentication token",
        "tags": [
          "Checks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Checks if the provided authentication token is valid.",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer <your-token-here>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token is valid"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "user123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "403": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
...
  },
...
}